Pass the full Configuration object to the sandbox.
